Fixed bug whereby using expired interactions will remove bot token (#67)
Bot would fail to use simple actions like `channel.send()` or `channel.sendTyping()` because the token was null.

It was very odd for that to happen because the bot works fine usually, this just happens out of the blue. After a little exploring, happened upon this thread: https://www.answeroverflow.com/m/989066624537600070

In fact, this matched the exact scenario, right before every logs of missing token, there was a 401 from discord. Specifically, this came from when the user would click "change image" on a character, but then simply don't respond. The bot will try to edit the message after 15 minutes, but by then, the interaction has already expired. As a result, attempting to editReply will cause the bot to get a 401, and in turn, reset the bot token for safety.

As a fix, we will not editReply if the interaction is expired, and to prevent future errors, ensured that the collectors would expire in 10 minutes instead of 15, if they wanted to do something with the interaction, after it expired. Further monitoring will be done to make sure there aren't any missing places.